created celery schedule to run our stats gathering nightly added crontab schedule to prod settings.
If you'd like to run the celery beat server (the thing that runs the scheduled jobs), this is handy:

python manage.py celery worker --loglevel=info -B

also shuffled around GENERATED_ASSETS_STORAGE in settings. we have a standard value for this in prod and /tmp/.. seems reasonable in dev.