Serve source view using WARC files if they exist.
This adds a Python application based on pywb that runs alongside our
Django app and serves WARC files. WARC files will be used if
link.asset.warc_capture points to a “.warc” or “.warc.gz” file. (Until
we add WARC capture, the only way that would happen is if you manually
edit the database.)

DEPLOYMENT NOTE:
This depends on forked versions of three packages that aren’t in PyPI —
pywb, warc-tools, and cdx_writer. Hopefully the changes will be merged
upstream and they’ll be added to PyPI. In the meantime they’re living
in my repo and included as git submodules.

So, when you update for the first time, you need to run:

git submodule init
git submodule update

to pull down the libraries.