Add bookmarklet v1
I've built version one of the bookmarklet. You can see it on the "create link"
page at /manage/create/ (on the "develop" branch of the repo).

The interface is simple: the user clicks the button while viewing a page, and
Perma will automatically archive that page and generate a preview as usual.

Here's how it works on the backend:

- The bookmarklet passes data to
/service/bookmarklet-create/?v=1&url=http://something-to-archive.com/

- The view associated with that URL passes the query string data to
/manage/create/?v=1&url=http://something-to-archive.com/

- Some JavaScript on /manage/create/ reads the query string, populates the
"create" field, and submits the form