Our content-length limit was blocking PDFs over 1MB. We want to check for PDF size as we stream. Removed the content-length validation for PDFs.
There might be a better way, more unified way to look at archives
as we dowload? It feels like content-length is unreliable (the
host can set that value to anything they like in the header), and
if we check for mime-type, we're playing whack-a-mole (like I did
in this patch) and again, we're relying on the host to set the
proper value for content-type.

fixes #830