Add CSP to replay.perma.cc
This PR adds a fairly "loose" Content Security Policy on replay.perma.cc.

The intended policy can be summarized as _"Everything's allowed within the `<iframe>`, as long as it's same-origin."_

The goal is to help prevent at browser level potential _"service worker leaks"_ (meaning: requests meant to be intercepted by the service worker that end up reaching the network)  that the playback software might not have accounted for yet.