Temporary workaround for mobile PDF intersitials
@kmukk reported that the mobile interstitial for PDFs isn't working as expected: nothing happens on click.

After looking into the issue, it appears that this might be due to the way we interact with replayweb.page in that context: most definitely TBD, and this will require a proper diagnosis.

In the meantime: since it appears that we can now render PDFs within `<replay-web-page>` on mobile, in concertation with @kmukk and @bensteinberg, we've decided to disable the interstitial for the time being.

Users _should_ therefore now see the PDFs on mobile instead of the interstitial, and still be able to download them if they so choose.